(ruby client example) add comment saying that link-initialize is loaded from `body`, not `head` (#107)
* add comment to index.html

The example has the link-initialize script loaded in the `<body>`, but it is not explicitly mentioned that it needs to be here (as opposed to `<head>`). When it is placed in the `<head>`, a cryptic error is raised:

```
plaid-link-initialize.js:2 Uncaught TypeError: Cannot read property 'appendChild' of null
```

Searching the issues for this error finds https://github.com/plaid/link/issues/67 which recommends moving the script tag into the `<body>`. 

I think there should be a comment in the example html file which mentions this requirement explicitly. Not everyone knows to search the issues page.

* add comment to go example